+++ src/pages/AdditionalInfoOAuth.tsx
@@ -2,10 +2,7 @@ import logo from '@/assets/images/logo.png';
 import { Button, ErrorMessage } from '@/components/atoms';
 import { InputForm } from '@/components/molecules';
 import { OAuthAPI, OAuthDomainType } from '@/services/auth';
-import {
-    additionalInfoSchema,
-    type AdditionalInfoSchema,
-} from '@/types/auth/schema';
+import { additionalInfoSchema, type AdditionalInfoSchema } from '@/types/auth';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm, type SubmitHandler } from 'react-hook-form';
 import { Navigate, useParams } from 'react-router';
